backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
pillow>=10.0.0
numpy>=1.26.0
reportlab>=4.0.0
python-multipart>=0.0.9
pydantic>=2.6.0
opencv-python-headless>=4.9.0.80
# Optional: only installed on non-Windows by default for convenience
# Remove markers if you want to install torch/torchvision on Windows
torch>=2.2.0; platform_system != "Windows"
torchvision>=0.17.0; platform_system != "Windows"
matplotlib>=3.8.0


